+++ src/main/java/com/br/alldreams/jupiter/banner/controller/BannerController.java
@@ -1,0 +1,17 @@
+/**
+ *
+ */
+package com.br.alldreams.jupiter.banner.controller;
+
+import org.springframework.stereotype.Controller;
+
+/**
+ * @author Jorge Demetrio
+ * @version 1.0.0
+ * @since 15 de jan de 2020 13:48:51
+ */
+
+@Controller
+public class BannerController {
+
+}
