+++ src/main/java/com/br/alldreams/jupiter/base/BaseController.java
@@ -8,6 +8,6 @@ package com.br.alldreams.jupiter.base;
  * @since 12 de jan de 2020 04:25:14
  * @version 1.0
  */
-public abstract class BaseController {
+public abstract class BaseController { // extends AbstractController {
 
 }
