+++ src/main/java/com/br/alldreams/jupiter/busca/controller/BuscaController.java
@@ -1,0 +1,19 @@
+/**
+ *
+ */
+package com.br.alldreams.jupiter.busca.controller;
+
+import org.springframework.stereotype.Controller;
+
+import com.br.alldreams.jupiter.base.BaseController;
+
+/**
+ * @author Jorge Demetrio
+ * @version 1.0.0
+ * @since 15 de jan de 2020 13:45:04
+ */
+
+@Controller
+public class BuscaController extends BaseController {
+
+}
