+++ src/main/java/com/br/alldreams/jupiter/conteudo/pagina/repository/model/StatusConteudoEnum.java
@@ -1,0 +1,27 @@
+/**
+ *
+ */
+package com.br.alldreams.jupiter.conteudo.pagina.repository.model;
+
+/**
+ * @author Jorge Demetrio
+ * @since 14 de jan de 2020 23:01:34
+ * @version 1.0
+ */
+public enum StatusConteudoEnum {
+	/**
+	 * Cotnteúdo deletado
+	 */
+	DELETADO,
+
+	/**
+	 * Conteduo publicado
+	 */
+	PUBLICADO,
+
+	/**
+	 * Não publicado
+	 */
+	NAO_PUBLICADO;
+
+}
