+++ src/main/java/com/br/alldreams/jupiter/conteudo/pagina/repository/model/StatusUsuarioEnum.java
@@ -1,0 +1,14 @@
+/**
+ *
+ */
+package com.br.alldreams.jupiter.conteudo.pagina.repository.model;
+
+/**
+ * @author Jorge Demetrio
+ * @since 15 de jan de 2020 00:31:52
+ * @version 1.0
+ */
+public enum StatusUsuarioEnum {
+
+	DELETADO, ATIVO, BLOQUEADO, EXPIRADO;
+}
