+++ src/main/java/com/br/alldreams/jupiter/enquete/controller/EnqueteController.java
@@ -1,0 +1,18 @@
+/**
+ *
+ */
+package com.br.alldreams.jupiter.enquete.controller;
+
+import org.springframework.stereotype.Controller;
+
+import com.br.alldreams.jupiter.base.BaseController;
+
+/**
+ * @author Jorge Demetrio
+ * @version 1.0.0
+ * @since 15 de jan de 2020 14:00:45
+ */
+@Controller
+public class EnqueteController extends BaseController {
+
+}
