--- src/main/java/com/br/alldreams/jupiter/materia/repository/model/StatusConteudoEnum.java
@@ -1,27 +1,0 @@
-/**
- *
- */
-package com.br.alldreams.jupiter.materia.repository.model;
-
-/**
- * @author Jorge Demetrio
- * @since 14 de jan de 2020 23:01:34
- * @version 1.0
- */
-public enum StatusConteudoEnum {
-	/**
-	 * Cotnteúdo deletado
-	 */
-	DELETADO,
-
-	/**
-	 * Conteduo publicado
-	 */
-	PUBLICADO,
-
-	/**
-	 * Não publicado
-	 */
-	NAO_PUBLICADO;
-
-}
