--- src/main/java/com/br/alldreams/jupiter/materia/repository/model/StatusUsuarioEnum.java
@@ -1,14 +1,0 @@
-/**
- *
- */
-package com.br.alldreams.jupiter.materia.repository.model;
-
-/**
- * @author Jorge Demetrio
- * @since 15 de jan de 2020 00:31:52
- * @version 1.0
- */
-public enum StatusUsuarioEnum {
-
-	DELETADO, ATIVO, BLOQUEADO, EXPIRADO;
-}
