--- src/main/java/com/br/alldreams/jupiter/materia/repository/model/TipoImagemEnum.java
@@ -1,13 +1,0 @@
-/**
- *
- */
-package com.br.alldreams.jupiter.materia.repository.model;
-
-/**
- * @author Jorge Demetrio
- * @since 14 de jan de 2020 23:29:23
- * @version 1.0
- */
-public enum TipoImagemEnum {
-	DESTAQUE, MATERIA, THUMB;
-}
