+++ src/main/java/com/br/alldreams/jupiter/usuario/controller/UsuarioController.java
@@ -1,0 +1,20 @@
+/**
+ *
+ */
+package com.br.alldreams.jupiter.usuario.controller;
+
+import org.springframework.stereotype.Controller;
+
+import com.br.alldreams.jupiter.base.BaseController;
+
+/**
+ * @author Jorge Demetrio
+ * @version 1.0.0
+ * @since 15 de jan de 2020 13:22:06
+ */
+@Controller
+public class UsuarioController extends BaseController {
+
+
+
+}
